requirements.txt
streamlit==1.24.0
pandas
python-dateutil
python-dotenv
simplegmail
langchain-groq==0.2.4
langchain==0.3.15
markitdown
Pillow<10 --only-binary=:all:
pyttsx3
SpeechRecognition
PyAudio
beautifulsoup4
requests
